+++ src/lineage/index/view.py
@@ -1,5 +1,7 @@
+from Products.CMFCore.utils import getToolByName
 from Products.Five.browser import BrowserView
 from plone.memoize.view import memoize_contextless
+from plone.uuid.interfaces import IUUID
 from zope.component import getUtility
 from zope.schema.interfaces import IVocabularyFactory
 
@@ -16,3 +18,14 @@ class ChildsiteView(BrowserView):
             return vocab.getTerm(key).title
         except LookupError:
             return key
+
+    def childsiteForContext(self, item):
+        """Returns the childsite UUID for a context object by looking up it's
+        brain in the catalog.
+        """
+        childsite = ''
+        cat = getToolByName(self.context, 'portal_catalog')
+        res = cat.searchResults(UID=IUUID(item), path='/')  # TODO: get portal root
+        if res:
+            childsite = res[0].childsite
+        return childsite